(#94)refactor: 검색 인풋창 컴포넌트 화 후, 필요한 곳에 import 하여 사용했습니다.
