+++ src/asset/sass/pages/searchPage/searchResultPage.scss
@@ -1,3 +1,8 @@
 .result-count {
   color: #ff8d1d;
+  margin-left: 1.5%;
+}
+
+.divider {
+  color: #000;
 }
